requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0

# GCP
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
google-cloud-aiplatform==1.65.0

# LangChain + Vertex AI
langchain==0.3.1
langchain-google-vertexai==2.0.0
langchain-google-community==2.0.1

# Twilio
twilio==9.3.0

# Forms
python-multipart==0.0.9
